% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download.de.data}
\alias{download.de.data}
\title{Download Germany data (DE)}
\usage{
download.de.data(by.federal.state = FALSE)
}
\value{
data frame with latest data
}
\description{
Download Germany data (DE)
}
