% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download.es.data}
\alias{download.es.data}
\title{Download world data from Spain's Ministerio de Sanidad}
\usage{
download.es.data(by.state = FALSE)
}
\value{
a list with data and source string
}
\description{
Download world data from Spain's Ministerio de Sanidad
}
