% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download.it.data}
\alias{download.it.data}
\title{Download world data from Italy's Protezione Civile}
\usage{
download.it.data(by.state = TRUE)
}
\value{
a list with data and source string
}
\description{
Download world data from Italy's Protezione Civile
}
