% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download.john.hopkins}
\alias{download.john.hopkins}
\title{Download world data from John Hopkin's University}
\usage{
download.john.hopkins()
}
\value{
a list with data and source string
}
\description{
Download world data from John Hopkin's University
}
