% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download.pt.data}
\alias{download.pt.data}
\title{Download world data from Portugal's Direção Geral da Saúde}
\usage{
download.pt.data()
}
\value{
a list with data and source string
}
\description{
Download world data from Portugal's Direção Geral da Saúde
}
