% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_names.R
\name{eu.convert.names}
\alias{eu.convert.names}
\title{Convert country names in EU CDC dataset}
\usage{
eu.convert.names(val)
}
\arguments{
\item{state}{what}
}
\value{

}
\description{
Convert country names in EU CDC dataset
}
