% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{it.nuts.codici}
\alias{it.nuts.codici}
\title{Italian mapping of NUTS_2 regions}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 21 rows and 3 columns.
}
\usage{
it.nuts.codici
}
\description{
Italian mapping of NUTS_2 regions
}
\examples{
head(it.nuts.codici)
}
\keyword{data}
