% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_names.R
\name{jh.convert.names}
\alias{jh.convert.names}
\title{Convert country names in John Hopkins dataset}
\usage{
jh.convert.names(val)
}
\arguments{
\item{state}{what}
}
\value{

}
\description{
Convert country names in John Hopkins dataset
}
