% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_map.R
\name{proper.cases}
\alias{proper.cases}
\title{Maps type of case}
\usage{
proper.cases(value, capitalize = FALSE, capitalize.all = FALSE)
}
\arguments{
\item{value}{one of 3, 'confirmed', 'death', 'all'}

\item{capitalize}{only first word}

\item{capitalize.all}{all words}
}
\value{
a formatted version of value
}
\description{
Maps type of case
}
\examples{
proper.cases(c('death', 'confirmed'))
proper.cases(c('death', 'confirmed'), capitalize = TRUE)
}
