% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_bank.R
\name{wb.indicator}
\alias{wb.indicator}
\title{Get World Bank indicators by 100k population}
\usage{
wb.indicator(dat, indicator, name, convert.names.fun = NULL)
}
\arguments{
\item{dat}{data frame with state column to be merged}

\item{indicator}{code of indicator}

\item{name}{name to replace on indicator}

\item{conver.names.fun}{function that receives country and maps to new name (present in state column of dat)}
}
\value{
data frame with world bank data and ratio do cases (and deaths) per the world bank indicator
}
\description{
Get World Bank indicators by 100k population
}
\examples{
wb.indicator('SH.MED.NUMW.P3')
}
